@@ -5,7 +5,7 @@ author: fedeoo <fedeoo.zf@gmail.com>
 homepage: https://github.com/fedeoo/flip_card/
 
 environment:
-  sdk: ">=2.0.0-dev.68.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
