@@ -9,9 +9,13 @@ enum FlipDirection {
 }
 
 class AnimationCard extends StatelessWidget {
-  AnimationCard({this.child, this.animation, this.direction});
+  AnimationCard({
+    required this.child,
+    required this.animation,
+    required this.direction,
+  });
 
-  final Widget child;
+  final Widget? child;
   final Animation<double> animation;
   final FlipDirection direction;
 
@@ -19,7 +23,7 @@ class AnimationCard extends StatelessWidget {
   Widget build(BuildContext context) {
     return AnimatedBuilder(
       animation: animation,
-      builder: (BuildContext context, Widget child) {
+      builder: (BuildContext context, Widget? child) {
         var transform = Matrix4.identity();
         transform.setEntry(3, 2, 0.001);
         if (direction == FlipDirection.VERTICAL) {
@@ -47,8 +51,8 @@ class FlipCard extends StatefulWidget {
   /// The amount of milliseconds a turn animation will take.
   final int speed;
   final FlipDirection direction;
-  final VoidCallback onFlip;
-  final BoolCallback onFlipDone;
+  final VoidCallback? onFlip;
+  final BoolCallback? onFlipDone;
 
   /// When enabled, the card will flip automatically when touched. This behavior
   /// can be disabled if this is not desired. To manually flip a card from your
@@ -75,16 +79,16 @@ class FlipCard extends StatefulWidget {
   ///```
   final bool flipOnTouch;
 
-  const FlipCard(
-      {Key key,
-      @required this.front,
-      @required this.back,
-      this.speed = 500,
-      this.onFlip,
-      this.onFlipDone,
-      this.direction = FlipDirection.HORIZONTAL,
-      this.flipOnTouch = true})
-      : super(key: key);
+  const FlipCard({
+    Key? key,
+    required this.front,
+    required this.back,
+    this.speed = 500,
+    this.onFlip,
+    this.onFlipDone,
+    this.direction = FlipDirection.HORIZONTAL,
+    this.flipOnTouch = true,
+  }) : super(key: key);
 
   @override
   State<StatefulWidget> createState() {
@@ -94,9 +98,9 @@ class FlipCard extends StatefulWidget {
 
 class FlipCardState extends State<FlipCard>
     with SingleTickerProviderStateMixin {
-  AnimationController controller;
-  Animation<double> _frontRotation;
-  Animation<double> _backRotation;
+  late AnimationController controller;
+  late Animation<double> _frontRotation;
+  late Animation<double> _backRotation;
 
   bool isFront = true;
 
@@ -134,14 +138,14 @@ class FlipCardState extends State<FlipCard>
     controller.addStatusListener((status) {
       if (status == AnimationStatus.completed ||
           status == AnimationStatus.dismissed) {
-        if (widget.onFlipDone != null) widget.onFlipDone(isFront);
+        if (widget.onFlipDone != null) widget.onFlipDone!(isFront);
       }
     });
   }
 
   void toggleCard() {
     if (widget.onFlip != null) {
-      widget.onFlip();
+      widget.onFlip!();
     }
     if (isFront) {
       controller.forward();
@@ -175,7 +179,7 @@ class FlipCardState extends State<FlipCard>
     return child;
   }
 
-  Widget _buildContent({@required bool front}) {
+  Widget _buildContent({required bool front}) {
     // pointer events that would reach the backside of the card should be
     // ignored
     return IgnorePointer(
